+++ contrib/install-osvvm.sh
@@ -13,23 +13,23 @@ if ! command -v tclsh &>/dev/null; then
   exit 1
 fi
 
-git_wrapper https://github.com/osvvm/OsvvmLibraries 2024.11
+git_wrapper https://github.com/osvvm/OsvvmLibraries 2025.02
 
-# Temporary workaround for analysis warnings
-(cd Scripts && patch -p1 -N) <<EOF
-diff --git a/VendorScripts_NVC.tcl b/VendorScripts_NVC.tcl
-index c9aee5443c10..31a8d4e259b6 100644
---- a/VendorScripts_NVC.tcl
-+++ b/VendorScripts_NVC.tcl
-@@ -175,7 +175,7 @@ proc vendor_analyze_vhdl {LibraryName FileName args} {
-   variable VhdlLibraryFullPath
-   variable NVC_WORKING_LIBRARY_PATH
- 
--  set  GlobalOptions [concat --std=\${VhdlShortVersion} -H 128m --work=\${LibraryName}:\${NVC_WORKING_LIBRARY_PATH}.\${VhdlShortVersion} {*}\${VHDL_RESOURCE_LIBRARY_PATHS}]
-+  set  GlobalOptions [concat --std=\${VhdlShortVersion} -H 128m --stderr=error --work=\${LibraryName}:\${NVC_WORKING_LIBRARY_PATH}.\${VhdlShortVersion} {*}\${VHDL_RESOURCE_LIBRARY_PATHS}]
-   set  AnalyzeOptions [concat {*}\${args} \${FileName}]
-   puts "nvc \${GlobalOptions} -a \$AnalyzeOptions"
-   if {[catch {exec \$nvc {*}\${GlobalOptions} -a {*}\$AnalyzeOptions} AnalyzeErrorMessage]} {
+# Temporary workaround for VHDL-2008 compile error
+(cd osvvm && patch -p1 -N) <<EOF
+diff --git a/osvvm.pro b/osvvm.pro
+index 442bf5f40c1d..a5e3c06838d0 100644
+--- a/osvvm.pro
++++ b/osvvm.pro
+@@ -93,7 +93,7 @@ analyze CoverageVendorApiPkg_\${::osvvm::FunctionalCoverageIntegratedInSimulator}
+
+ analyze TranscriptPkg.vhd
+
+-if {\$::osvvm::Support2019FilePath} {
++if {\$::osvvm::Support2019FilePath && \$::osvvm::VhdlVersion >= 2019} {
+   analyze FileLinePathPkg.vhd
+ } else {
+   analyze deprecated/FileLinePathPkg_c.vhd
 EOF
 
 [ -d $NVC_INSTALL_DEST ] || mkdir -p $NVC_INSTALL_DEST